requirements.txt
sympy>=1.12
mpmath>=1.3.0
numpy>=1.26
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
sqlalchemy>=2.0
antlr4-python3-runtime==4.11
httpx>=0.27
python-dotenv>=1.0
structlog>=24.1
pytest>=8.0
pytest-asyncio>=0.23
